move precompile
